Refactor angles
